fixed a bug in managing keys
when selector used its keypress (of enter) it left
behind a keyup of enter, which is then passed into
the editor (for example when selecting a tag to
edit and editing it) which caused the editor to close
immediately. Now we have something in place that lets
you cancel a future element (which you know it may
lead to something like this). The function is called
cancel_next.

BTW, selector was missing the my_event stuff