tag editor is done
the json does not have the x.item.value and x.tag.value
so those had to change.

there is one bug which I have not found, probably due
to a missing x.(item|tag) which I have not yet changed.
To reproduce this bug do the following:

rake db:gitdb:drop db:gitdb:create db:gitdb:seed
http://127.0.0.1:3000/in

yank the first element, then press j and an exception will say that the
element is not found